Cleaned kyoto data
Cleaned kyoto data with temperature and precipitation with the metric unit.
(I changed the name to make it easy to understand. The standard version of csv file from the website may have a different column name)
The pdf version is just for visualization and helps us better understand the data.